+++ gamble/models/__init__.py
@@ -7,3 +7,4 @@ models submodule for gamble
 from gamble.models.dice import *
 from gamble.models.cards import *
 from gamble.models.golf import *
+from gamble.models.BlackJack import *
